CASMCMS-9421: Remove CPS/DVS rootfs support; Add rootfs provider validation to session template validation endpoint
